Fix div-by-zero

+++ src/main/java/org/metamechanists/quaptics/implementation/blocks/attachments/ProgressBlock.java
@@ -24,6 +24,6 @@ public interface ProgressBlock {
 
         final double progress = getProgress(location.get());
         final double maxTime = group.getBlock().getSettings().getTimePerItem();
-        return Lore.progressBar((int) progress, (int) maxTime, "&a", "&7");
+        return Lore.progressBar(progress, maxTime, "&a", "&7");
     }
 }
